readyset-e2e-tests: Don't run window function test against MySQL 5.7
It doesn't support them and just raises a syntax error when trying to
compare our results to upstream.

We also use slightly less magic in the macros in order to make it easier
(and avoid adding even more magic) to set a different upstream tag
(`mysql8_upstream`) for the window function cases while leaving the
normal aggregates alone. So this means we'll have to add another variant
to the non-window-function macro if we ever want to include value checks
in any of the non-window-function tests.

Change-Id: Iddcb1ef7135913e7c6b274e646247a1b6a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11017
Tested-by: Buildkite CI
Reviewed-by: Ron Hough <ron@readyset.io>
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>