Test all columns in SHOW CACHES
We already had integration tests that exercise the SHOW CACHES statement
through MySQL and PostgreSQL, but because the test setup did not provide
a MetricsHandle, SHOW CACHES omitted the count column.

Provide the handle to the integration tests so they see all of the
columns SHOW CACHES can return.

The first version of this patch mistakenly changed the BackendBuilder
used by everything.  This version only modifies the BackendBuilder in
test code.

Change-Id: Iebe301893d4c35ef6e0243db0dfca7250638fa1b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11029
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>