readyset-server: Allow computed local predicates
This allows more predicates referencing a single table within join `ON`
clauses to be treated as local to that table, which causes them to be
turned into filters between the table and the join.

For example, this was previously unsupported because the length
predicate became a "global" predicate which is unsupported for the join:

```sql
SELECT * FROM t LEFT JOIN v ON t.id = v.id AND length(v.field) > 10;
```

This pattern is not yet allowed *everywhere*, but allows predicates
which are binary expressions only referencing a single table, even if
they contain computed expressions on columns in that table.

Change-Id: I5a8ef482ee5a7e670de6b98e4713abad6a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11069
Tested-by: Buildkite CI
Reviewed-by: Vassili Zarouba <vassili@readyset.io>