readyset-server: Add regression test for compaction timeout
This was originally written mostly to verify which timeout was being hit
when compaction exceeded it, but may be used in the future to verify
improvements (such as allowing compaction to exceed the timeout or
making it a background task).

To aid in manual/end-to-end testing of this failure, we also make it
possible to set failure points at start up in readyset by adding a
`FailScenario` to `main`. Now, we can set the `FAILPOINTS` environment
variable.

AI-Use: level:3
Change-Id: Ifa85dd0e265002ad7f795045e88987fd6a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11084
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>