dataflow-state: Don't block table requests on snapshot compaction
After leaving snapshot mode we now fire off manual compactions for all
column families without joining the threads, allowing the domain to
continue processing other requests while compaction finishes in the
background. This also means we don't have on giant table request waiting
for compaction to finish, which would cause that request itself to
time out.

Removing this wait is presumed safe based on the belief that we were
previously only waiting on compaction to avoid performance degradation
(e.g. huge L0 in RocksDB) from streaming replication starting when they
are still compacting, and there was no correctness problem that waiting
was solving.

However, we wait for compaction to finish after all tables are
snapshotted anyway: we were calling `wait_for_all_tables_to_compact` at
the end of `PostgresReplicator::snapshot_to_noria` already, and now do
so in `MysqlReplicator` as well. So we still have that performance
safeguard; but now we don't hold up domains on it, we just let the
replicator poll for it after triggering the compaction.

This means we shouldn't hit `TABLE_REQUEST_TIMEOUT_SECONDS` in practice,
at least for compactions, unless there's a bug.

AI-Use: level:2
Change-Id: I9e1ff9c391d3214749d15441bc8c35976a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11090
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>