replicators: Report server UUID when registering as a replica
MySQL 5.7 and some earlier versions of MySQL 8.0 have a bug where
it will report a malformed packet when executing SHOW SLAVE HOSTS or
SHOW REPLICAS. The issue is fixed in MySQL 8.0.28. In summary, the
affected versions will attempt to report the server UUID, if this is
not set when registering as a replica via SET @slave_uuid=<uuid> the
command will fail.

The fix is to add a server UUID parameter and use it or generate a
random UUIDv4 when registering as a replica.

Closes: REA-6238
Fixes: #1575

Release-Note-Core: Added the option `--replication-server-uuid` to
  set a server UUID in MySQL. This UUID will be reported in
  `SHOW REPLICAS`.
Change-Id: I75df8080ab25691600b207b999ddb1f4062cd07c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11105
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI