readyset-adapter: Support EXPLAIN CREATE [DEEP|SHALLOW] CACHE
This patch adds a shallow cache support check to EXPLAIN CREATE CACHE.

Shallow cache support is determined by whether or not we can prepare the
query on the upstream.

If the DEEP or SHALLOW keyword is added, we will limit our support check
to the requested variant.  Otherwise, we report support according to the
current CacheMode.

If we're running with CacheMode::DeepThenShallow, we'll say a query is
supported if we can deep cache it or shallow cache it.

Also include the reason why we don't support caching a query.  For deep
caches, the reason now matches what is shown in the SHOW PROXIED QUERIES
output.  For shallow caches, the reason is the error provided by the
upstream database.

Fixes: REA-6169
Change-Id: I88a6ffeb8c1e70c292263fa16989d5c7d12ded20
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11160
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>