readyset-adapter: Implement shallow cache fallback
When a CREATE CACHE statement is provided without a DEEP or SHALLOW
modifier, Readyset will now infer the requested cache type based on the
value provided to the --cache-mode argument:

  --cache-mode deep

      Readyset will only create deep caches.

  --cache-mode shallow

      Readyset will only create shallow caches.

  --cache-mode deep-then-shallow

      Readyset will first try to create deep caches, falling back to
      creating shallow caches if deep caching is not supported.

Providing DEEP or SHALLOW to the CREATE CACHE statement will override
this behavior and create the requested cache type.

Since the current default is --cache-mode deep, no behavior is changed
so long as the argument is omitted.

Addresses: REA-6151
Change-Id: I7d30961f8b52fac3db65a91eac0c1e0ceed0a2b5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11163
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI