readyset-sql-passes: Add full parameterization rewrite
Readyset currently selects cache key parameters based on the existing
autoparameterization process, which attempts to limit placeholder
positions to the subset of positions expected to work with deep caching.

We want to support a lot more placeholder positions with shallow
caching, so this patch introduces an alternate process which fully
replaces all literals in the query with placeholders, extracting all
literal values as parameters usable for cache keys.

This isn't hooked up to anything yet, but this alternate process can be
selected by passing ParameterizeMode::Full to rewrite_equivalent.

Change-Id: I431ac7b3aad52d92a58a91991d2955c4d908a8f6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11175
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>