readyset-adapter: Fully-parameterize shallow caches
This patch should drastically increase query coverage for shallow
caches.

When creating shallow caches, first replace all query literals with
placeholders.  Note that this form's query_id is likely different from
from the one calculated from its typical autoparameterized form.

When receiving a query, we now similarly rewrite it by extracting all
literals as cache parameters, replacing the literals with placeholders.
If a matching shallow cache is found, we proceed with query processing,
allowing the adapter to then utilize that shallow cache.

If a shallow cache is not found, we rollback and instead proceed with
the typical autoparameterization rewrite.

Change-Id: I7d875d50d590c41101aef0015823eb79fd19e360
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11176
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI