readyset-server: Add back some Antithesis reachability assertions
Expand some Antithesis reachability assertions to have a static callsite
per enum variant (for `MirNodeInner` and `FunctionExpr` when used in
post lookups).

Our previous de-noising (Idbfd30c6, I0aa1b7c8, I67a0d2a8) left us with
just *one* reachability assertion for each of the 3 main places
(dataflow node types, built-in functions, and aggregation functions),
because `antithesis_sdk` relies on static callsite instrumentation to
say "there was an assertion on line 42 that wasn't hit". So for example,
even if we only ever lowered a base node, we would get one reachability
passed assertion and we wouldn't be alerted that we didn't hit the other
node types.

Now that we have the weekly rollup (I386e4c3d), the extra noise won't be
as much of a problem (we can even suppress this assert on nightly
reports). So we go back to one static callsite per node type for
dataflow and one per aggregation function.

For now, I'm leaving the built-in functions as they are, since they are
lower stakes, more numerous, and don't already have an enum to easily...
enumerate. We might re-enable all of them individually if the rollup
works well for us (or we improve our test efficiency and can hit
everything every nightly).

Addresses: REA-6276
AI-Use: level:v
Change-Id: If819f6f3a24da3ba7254ee4c1410ab5c6a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11213
Reviewed-by: Johnathan Davis <jcd@readyset.io>
Tested-by: Buildkite CI