readyset-adapter: Rework EXPLAIN CREATE CACHE for shallow cache ASTs
The deep cache and shallow cache support checks each use a different AST
due to rewrites alone, and this divergence will become even greater when
shallow caches use sqlpaser AST types instead of our own AST types.

The explain code previously accepted only the Readyset-rewritten AST
suitable for deep caching, which is right some of the time, but not
always.  Rework the explain code to be aware of both deep and shallow
AST representations and use whichever representation is appropriate
during the support check.

Some of the patch is cleanup around the decision of whether to perform a
dry run, which uncovered and fixed a bug where we wouldn't dry run a
query when --cache-mode shallow and EXPLAIN CREATE DEEP was requested.

Change-Id: I8f450cf89270b97ea476156f7feb7b54a8799d5c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11244
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>